Move decorators & some utils, and add multidim
