+++ neurodsp/tests/test_utils_decorators.py
@@ -7,3 +7,6 @@ from neurodsp.sim.decorators import *
 
 def test_normalize():
     pass
+
+def test_multidim():
+    pass
